Modify 판매 상품, 구매 상품 price double 타입

+++ src/main/java/com/metalancer/backend/orders/domain/DaySalesReport.java
@@ -8,11 +8,11 @@ public class DaySalesReport {
 
     private final String day;
     private final int totalPrice;
-    private final int totalPriceUSD;
+    private final Double totalPriceUSD;
     private final int salesCnt;
 
     @Builder
-    public DaySalesReport(String day, Integer totalPrice, Integer totalPriceUSD, int salesCnt) {
+    public DaySalesReport(String day, Integer totalPrice, Double totalPriceUSD, int salesCnt) {
         this.day = day;
         this.totalPrice = totalPrice != null ? totalPrice : 0;
         this.totalPriceUSD = totalPriceUSD != null ? totalPriceUSD : 0;
